Adding additional factories and documentation.
