+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/component/multi/KlMultiVersionPaneFactory.java
@@ -2,8 +2,8 @@ package dev.ikm.komet.layout.component.multi;
 
 import dev.ikm.komet.framework.observable.ObservableVersion;
 import dev.ikm.komet.framework.view.ObservableView;
-import dev.ikm.komet.layout.KlFactory;
 import dev.ikm.komet.layout.KlVersionType;
+import dev.ikm.komet.layout.KlWidgetFactory;
 import dev.ikm.komet.layout.component.KlComponentPane;
 import dev.ikm.komet.preferences.KometPreferences;
 import org.eclipse.collections.api.list.ImmutableList;
@@ -14,7 +14,7 @@ import org.eclipse.collections.api.list.ImmutableList;
  * @param <OV> the type parameter that extends {@link ObservableVersion}, representing the specific version type
  *             that this factory interacts with.
  */
-public interface KlMultiVersionPaneFactory<OV extends ObservableVersion> extends KlFactory, KlVersionType<OV> {
+public interface KlMultiVersionPaneFactory<OV extends ObservableVersion> extends KlWidgetFactory, KlVersionType<OV> {
 
     /**
      * Creates an instance of {@link KlMultiVersionPane} tailored to the specified observable versions, observable view,
