+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/component/multi/package-info.java
@@ -1,0 +1,25 @@
+/**
+ * The `dev.ikm.komet.layout.component.multi` package focuses on managing and organizing
+ * multi-component layouts (Multiple Concept, Semantics, Pattern, and Stamp KlWidgets) within
+ * the Komet framework. These components enable the
+ * representation and interaction of multiple related elements or views in a unified
+ * and cohesive manner.
+ *
+ * <p>
+ * Key functionalities include:
+ * <ul>
+ *   <li>Providing tools and abstractions for designing layouts with multiple components.</li>
+ *   <li>Supporting the interaction and synchronization of multiple elements within a shared
+ *       layout context.</li>
+ *   <li>Enabling customization and configuration of multi-component views to suit domain-specific
+ *       requirements.</li>
+ * </ul>
+ * </p>
+ *
+ * <p>
+ * This package is particularly useful for developers building complex user interfaces that require
+ * a high degree of interactivity and flexibility, where multiple components must be displayed
+ * and managed simultaneously.
+ * </p>
+ */
+package dev.ikm.komet.layout.component.multi;
