+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/component/package-info.java
@@ -1,0 +1,27 @@
+/**
+ * TODO: Consider renaming this package to "container", or moving inside the container package
+ * and naming it entity...
+ * The `dev.ikm.komet.layout.component` package focuses on defining and managing
+ * KlWidgets that present components (Concepts, Patterns, Semantics, and Stamps)
+ * within the Komet framework. These components are responsible
+ * for structuring and presenting data in a modular and visually organized manner,
+ * tailored to the needs of the application and its users.
+ *
+ * <p>
+ * Key functionalities include:
+ * <ul>
+ *   <li>Providing base abstractions and utilities for creating reusable and
+ *       customizable layout components.</li>
+ *   <li>Supporting component-based organization of data while ensuring flexibility
+ *       in layout design and data presentation.</li>
+ *   <li>Integrating with other parts of the Komet UI system for seamless user experiences.</li>
+ * </ul>
+ * </p>
+ *
+ * <p>
+ * This package is designed to be used by developers implementing UI layouts and
+ * workflows, ensuring a consistent and structured approach to constructing
+ * interactive user interfaces.
+ * </p>
+ *
+ */package dev.ikm.komet.layout.component;
