+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/component/version/KlVersionPaneFactory.java
@@ -2,11 +2,11 @@ package dev.ikm.komet.layout.component.version;
 
 import dev.ikm.komet.framework.observable.ObservableVersion;
 import dev.ikm.komet.framework.view.ObservableView;
-import dev.ikm.komet.layout.KlFactory;
 import dev.ikm.komet.layout.KlVersionType;
+import dev.ikm.komet.layout.KlWidgetFactory;
 import dev.ikm.komet.preferences.KometPreferences;
 
-public interface KlVersionPaneFactory <OV extends ObservableVersion> extends KlFactory, KlVersionType<OV> {
+public interface KlVersionPaneFactory <OV extends ObservableVersion> extends KlWidgetFactory, KlVersionType<OV> {
 
     KlVersionPane<OV> create(OV observableVersion,
                            ObservableView observableView,
