+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/component/version/field/package-info.java
@@ -1,4 +1,27 @@
 package dev.ikm.komet.layout.component.version.field;
 /**
-
+ * This package, `dev.ikm.komet.layout.component.version.field`, contains classes and components
+ * related to the management, organization, and presentation of versioned fields in the context
+ * of layout or UI components. It is designed to support and enhance the visualization and
+ * interaction with versioned data fields in a component-based architecture.
+ *
+ * Key Features:
+ * - Handles version-based field presentations, ensuring that UI components can dynamically
+ *   reflect and interact with versioned data elements.
+ * - Provides mechanisms to structure and layout versioned fields in a meaningful and consistent way.
+ * - Supports customization and extension to cater to domain-specific requirements for versioned data.
+ *
+ * Typical Use Cases:
+ * - Rendering components in UI layouts where version fields need to be displayed or managed.
+ * - Encapsulating logic for managing versioned states for individual fields, with potential support
+ *   for editing, interacting, or navigating through historical versions of data fields.
+ *
+ * Integration Notes:
+ * - Classes in this package are designed to work seamlessly with other layout and UI frameworks
+ *   in the `komet` application.
+ * - Expected to be utilized by developers implementing layouts or UI components involving versions,
+ *   enabling a modular approach to field version management.
+ *
+ * This package plays a critical role in ensuring that the interaction with versioned components
+ * remains intuitive and developer-friendly.
  */
