+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/component/version/package-info.java
@@ -1,0 +1,36 @@
+/**
+ * This package, `dev.ikm.komet.layout.component.version`, is responsible for managing
+ * and organizing the layout of entity version KlWidgets within the Komet Knowledge Layout framework.
+ *
+ * Purpose:
+ * - The primary focus of this package is to provide service contracts for KlWidgets that present
+ *   entity versions for use in UI layout components.
+ * - It enables developers to work with components whose states evolve over time,
+ *   ensuring that versioning is handled in a structured and consistent manner.
+ *
+ * Key Responsibilities:
+ * - Encapsulating logic for managing versioned states of components and their relationships.
+ * - Facilitating the display, organization, and navigation of versioned data in a UI context.
+ * - Providing utilities or abstractions for interacting with versioned elements in a
+ *   layout or model-driven workflow.
+ *
+ * Use Cases:
+ * - Supporting the creation and management of layouts that involve version-sensitive
+ *   components.
+ * - Ensuring version tracking and presentation are seamlessly integrated into UI workflows.
+ * - Allowing users or systems to interact with historical, current, or future states of
+ *   layout components.
+ *
+ * Integration Notes:
+ * - This package is part of the larger Komet framework and is designed to work with
+ *   related layout and UI components.
+ * - It plays a critical role in interfacing with other versioning-related modules or
+ *   services to ensure consistency and cohesion across the application.
+ *
+ * Design Goals:
+ * - To provide a modular, reusable set of tools for managing versioned components.
+ * - To simplify the process of integrating version tracking and navigation into layouts.
+ * - To ensure the maintainability and scalability of versioned UI components in
+ *   diverse use cases.
+ */
+package dev.ikm.komet.layout.component.version;
