+++ knowledge-layout/src/main/java/dev/ikm/komet/layout/window/package-info.java
@@ -1,0 +1,38 @@
+/**
+ * Provides classes and interfaces for handling window layout management within the Komet application.
+ *
+ * <p>The {@code dev.ikm.komet.layout.window} package is responsible for managing and organizing
+ * the application's windowing functionality. It provides tools and structures to define, manage,
+ * and interact with dynamic window layouts.</p>
+ *
+ * <h2>Core Responsibilities</h2>
+ * <ul>
+ *   <li>Define and manage window layout configurations.</li>
+ *   <li>Support user interactions with windows, such as resizing, dragging, and maximizing/minimizing.</li>
+ *   <li>Persist and reload window layout states as part of the application's lifecycle.</li>
+ *   <li>Facilitate integration of window layouts with other UI-related components or services.</li>
+ * </ul>
+ *
+ * <h2>Design Considerations</h2>
+ * <ul>
+ *   <li><b>Decoupled Components:</b> Ensure modular and independent classes for reuse and maintainability.</li>
+ *   <li><b>Clear Naming Practices:</b> Use descriptive names for classes and methods to explicitly
+ *       describe their purposes within the layout management system.</li>
+ *   <li><b>Performance Optimization:</b> Design window layout management with efficient
+ *       runtime performance in mind to ensure smooth user experiences.</li>
+ *   <li><b>Error Handling:</b> Provide robust support for edge cases, such as unexpected user interactions
+ *       and missing layout data.</li>
+ * </ul>
+ *
+ * <h2>Potential Use Cases</h2>
+ * <ul>
+ *   <li>Loading previously saved window layouts during application startup.</li>
+ *   <li>Allowing users to dynamically adjust window layouts through interactions such as drag, drop, or resize.</li>
+ *   <li>Defining custom layout structures programmatically or with templates.</li>
+ * </ul>
+ *
+ * <p>The package is central to the Komet application's user interface infrastructure. Its design and
+ * implementation ensure consistent, efficient, and customizable layout management for end-users.</p>
+ *
+ */
+package dev.ikm.komet.layout.window;
